refactor Procfile
