+++ utils.py
@@ -1,12 +1,12 @@
 from typing import Union
 from requests import User
-from jose import JWTError, jwt
+from jose import jwt
 from sqlalchemy.orm import Session
 from email_validate import validate
-from db.models import User as UserModel
+from models import User as UserModel
 from passlib.context import CryptContext
 from datetime import datetime, timedelta
-from fastapi import HTTPException, Depends
+from fastapi import HTTPException
 from fastapi.security import OAuth2PasswordRequestForm
 
 SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
